pyproject.toml
[project]
name = "multivision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.26.0",
    "channels>=4.0",
    "daphne>=4.0",
    "diffusers>=0.25.0",
    "django>=5.0",
    "django-cors-headers>=4.3",
    "djangorestframework>=3.14",
    "mediapipe>=0.10.33",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.0.0",
    "python-dotenv==1.0.0",
    "safetensors>=0.4.0",
    "timm>=1.0.26",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.36.0",
]
